update generator code

+++ TanyaPortal/src/main/java/com/srct/service/tanya/portal/vo/admin/tanya/DiscountInfoEntityVO.java
@@ -10,8 +10,8 @@ import java.util.Date;
 * 实体类
 * 根据需要删减无效参数
  * 
-* @author: Sharp
-* @create: 2019/01/28
+* @author: srct
+* @create: 2019/01/29
  **/
 @Data
 public class DiscountInfoEntityVO {
@@ -56,5 +56,5 @@ public class DiscountInfoEntityVO {
     private Date confirmAt;
  
     @ApiModelProperty(value = "有效性")
-    private Integer valid;
+    private Byte valid;
 }
