+++ TanyaPortal/src/main/java/com/srct/service/tanya/portal/vo/admin/tanya/MerchantInfoEntityVO.java
@@ -10,8 +10,8 @@ import java.util.Date;
 * 实体类
 * 根据需要删减无效参数
  * 
-* @author: Sharp
-* @create: 2019/01/28
+* @author: srct
+* @create: 2019/01/29
  **/
 @Data
 public class MerchantInfoEntityVO {
@@ -44,5 +44,5 @@ public class MerchantInfoEntityVO {
     private Date lastAt;
  
     @ApiModelProperty(value = "有效性")
-    private Integer valid;
+    private Byte valid;
 }
